+++ source/js/common/form/validationHelpers/index.js
@@ -1,5 +1,7 @@
 import radValidateHoc from './radValidateHoc';
+import validation from './validation';
 
 export {
 	radValidateHoc,
+  validation
 };
