+++ source/js/common/form/validationHelpers/validation.js
@@ -1,11 +1,16 @@
-import {validateHelper} from 'modul-helpers';
+import { validateHelper } from 'modul-helpers';
 
-export const isEmail = message => ({field}) => {
+const isEmail = message => ({ field }) => {
   const isValid = validateHelper.validEmail(field.value);
   return [isValid, message];
 };
 
-export const isRequired = message => ({field}) => {
+const isRequired = message => ({ field }) => {
   const isValid = !validateHelper.isEmpty(field.value);
   return [isValid, message];
+};
+
+export {
+  isEmail,
+  isRequired,
 };
