+++ source/js/modules/account/pages/ProfilePage.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { observer, inject } from 'mobx-react';
+
+@inject('profileStore')
+@observer
+class ProfileContainer extends React.Component {
+
+  render() {
+    return (
+      <div>
+      </div>
+    );
+  }
+}
+
+export default ProfileContainer;
